+++ arrow-udf/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support type alias `numeric` and `jsonb` in `#[function]`.
+- Support `varchar[]` and `bytea[]` as arguments.
+
 ## [0.1.1] - 2024-01-17
 
 ### Fixed